feat - Added assign to me button
One of the most common scenarios I found while using Taiga is that I
assign User Stories or Issues to myself, looking at other tools such as
Gitlab it comes with a quick 'Assign to me' button for merge requests.

In this pull request I have added a 'Assign to me' button with the same
styling as the 'Add Watchers' button. This shows in both User Stories
and Issues.